+++ tfexec/options.go
@@ -204,14 +204,6 @@ func Out(path string) *OutOption {
 	return &OutOption{path}
 }
 
-type OutputNameOption struct {
-	name string
-}
-
-func OutputName(name string) *OutputNameOption {
-	return &OutputNameOption{name}
-}
-
 type ParallelismOption struct {
 	parallelism int
 }